+++ lib/Main.dart
@@ -1,4 +1,4 @@
 import 'package:flutter/cupertino.dart';
-import 'package:islami/splash.dart';
+import 'package:islami/Splash.dart';
 
 void main() => runApp(SplashWidget());
