+++ lib/quran/SuraLink.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:islami/sura_content/sura_content_arguments.dart';
-import 'package:islami/sura_content/sura_content.dart';
+import 'package:islami/sura_content/SuraContentArguments.dart';
+import 'package:islami/sura_content/SuraContent.dart';
 class SuraLink extends StatelessWidget {
   const SuraLink ({Key? key, required this.suraNumber, required this.text, this.decoration, this.suraName}) : super(key: key);
